AppStore and Google Play profitable freemium app analysis
Analysis of which free apps would the most profitable for a company whose main area of business is developing free apps. 

Observation, data cleaning and analysis specifically on the ratio of app availability / downloads for each genre in the App Store and Google Play Store. Described the process step by step aiming the analysis to Director and above level executives that might not have the technical knowledge to understand the code but that need this strategic data to keep the business growing. At the end of the project, I draw my conclusions and give a sample of the best apps in each store for the development team to have a starting point.